requirements.txt
alembic==1.11.1
Babel==2.9.0
blinker==1.6.2
Bootstrap-Flask==2.2.0
click==8.1.3
colorama==0.4.6
distlib==0.3.6
ecdsa==0.18.0
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
phonenumbers==8.13.17
platformdirs==3.8.1
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.0
PyJWT==2.8.0
python-dateutil==2.6.0
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.17
typing_extensions==4.7.1
virtualenv==20.23.1
Werkzeug==2.3.6
WTForms==3.0.1